+++ tests/algorithms/x/Java/neural_network/activation_functions/softplus.java
@@ -1,31 +1,31 @@
 public class Main {
 
     static double ln(double x) {
-        if ((double)(x) <= 0.0) {
+        if ((double)(x) <= (double)(0.0)) {
             throw new RuntimeException(String.valueOf("ln domain error"));
         }
-        double y_1 = ((double)(x) - 1.0) / ((double)(x) + 1.0);
-        double y2_1 = y_1 * y_1;
-        double term_1 = y_1;
-        double sum_1 = 0.0;
+        double y_1 = (double)((double)(((double)(x) - (double)(1.0))) / (double)(((double)(x) + (double)(1.0))));
+        double y2_1 = (double)((double)(y_1) * (double)(y_1));
+        double term_1 = (double)(y_1);
+        double sum_1 = (double)(0.0);
         long k_1 = 0L;
-        while ((long)(k_1) < (long)(10)) {
-            double denom_1 = ((Number)(((long)((long)(2) * (long)(k_1)) + (long)(1)))).doubleValue();
-            sum_1 = sum_1 + term_1 / denom_1;
-            term_1 = term_1 * y2_1;
-            k_1 = (long)((long)(k_1) + (long)(1));
+        while ((long)(k_1) < 10L) {
+            double denom_1 = (double)(((Number)(((long)(2L * (long)(k_1)) + 1L))).doubleValue());
+            sum_1 = (double)((double)(sum_1) + (double)((double)(term_1) / (double)(denom_1)));
+            term_1 = (double)((double)(term_1) * (double)(y2_1));
+            k_1 = (long)((long)(k_1) + 1L);
         }
-        return 2.0 * sum_1;
+        return (double)(2.0) * (double)(sum_1);
     }
 
     static double exp(double x) {
-        double term_2 = 1.0;
-        double sum_3 = 1.0;
+        double term_2 = (double)(1.0);
+        double sum_3 = (double)(1.0);
         long n_1 = 1L;
-        while ((long)(n_1) < (long)(20)) {
-            term_2 = term_2 * (double)(x) / (((Number)(n_1)).doubleValue());
-            sum_3 = sum_3 + term_2;
-            n_1 = (long)((long)(n_1) + (long)(1));
+        while ((long)(n_1) < 20L) {
+            term_2 = (double)((double)((double)(term_2) * (double)(x)) / (double)((((Number)(n_1)).doubleValue())));
+            sum_3 = (double)((double)(sum_3) + (double)(term_2));
+            n_1 = (long)((long)(n_1) + 1L);
         }
         return sum_3;
     }
@@ -35,9 +35,9 @@ public class Main {
         long i_1 = 0L;
         while ((long)(i_1) < (long)(vector.length)) {
             double x_1 = (double)(vector[(int)((long)(i_1))]);
-            double value_1 = (double)(ln(1.0 + (double)(exp((double)(x_1)))));
+            double value_1 = (double)(ln((double)((double)(1.0) + (double)(exp((double)(x_1))))));
             result = ((double[])(appendDouble(result, (double)(value_1))));
-            i_1 = (long)((long)(i_1) + (long)(1));
+            i_1 = (long)((long)(i_1) + 1L);
         }
         return result;
     }
