+++ tests/algorithms/x/Java/other/nested_brackets.java
@@ -4,9 +4,9 @@ public class Main {
     static String[] slice_without_last(String[] xs) {
         String[] res = ((String[])(new String[]{}));
         long i_1 = 0L;
-        while ((long)(i_1) < (long)((long)(xs.length) - (long)(1))) {
+        while ((long)(i_1) < (long)((long)(xs.length) - 1L)) {
             res = ((String[])(java.util.stream.Stream.concat(java.util.Arrays.stream(res), java.util.stream.Stream.of(xs[(int)((long)(i_1))])).toArray(String[]::new)));
-            i_1 = (long)((long)(i_1) + (long)(1));
+            i_1 = (long)((long)(i_1) + 1L);
         }
         return res;
     }
@@ -15,22 +15,22 @@ public class Main {
         String[] stack = ((String[])(new String[]{}));
         long i_3 = 0L;
         while ((long)(i_3) < (long)(_runeLen(s))) {
-            String symbol_1 = _substr(s, (int)((long)(i_3)), (int)((long)((long)(i_3) + (long)(1))));
+            String symbol_1 = _substr(s, (int)((long)(i_3)), (int)((long)((long)(i_3) + 1L)));
             if (OPEN_TO_CLOSED.containsKey(symbol_1)) {
                 stack = ((String[])(java.util.stream.Stream.concat(java.util.Arrays.stream(stack), java.util.stream.Stream.of(symbol_1)).toArray(String[]::new)));
             } else             if ((symbol_1.equals(")")) || (symbol_1.equals("]")) || (symbol_1.equals("}"))) {
-                if ((long)(stack.length) == (long)(0)) {
+                if ((long)(stack.length) == 0L) {
                     return false;
                 }
-                String top_1 = stack[(int)((long)((long)(stack.length) - (long)(1)))];
+                String top_1 = stack[(int)((long)((long)(stack.length) - 1L))];
                 if (!(((String)(OPEN_TO_CLOSED).get(top_1)).equals(symbol_1))) {
                     return false;
                 }
                 stack = ((String[])(slice_without_last(((String[])(stack)))));
             }
-            i_3 = (long)((long)(i_3) + (long)(1));
+            i_3 = (long)((long)(i_3) + 1L);
         }
-        return (long)(stack.length) == (long)(0);
+        return (long)(stack.length) == 0L;
     }
 
     static void main() {
